fix warnings, cleanup

+++ demos/piano/soloud_basicwave.cpp
@@ -41,7 +41,7 @@ namespace SoLoud
 			case Basicwave::SINE:
 				for (i = 0; i < aSamples; i++)
 				{
-					aBuffer[i] = sin(mParent->mFreq * mOffset * M_PI * 2);
+					aBuffer[i] = (float)sin(mParent->mFreq * mOffset * M_PI * 2);
 					mOffset++;
 				}
 				break;
@@ -62,14 +62,14 @@ namespace SoLoud
 			case Basicwave::SQUARE:
 				for (i = 0; i < aSamples; i++)
 				{
-					aBuffer[i] = (fmod(mParent->mFreq * mOffset, 1) > 0.5) ? -1 : 1;
+					aBuffer[i] = (fmod(mParent->mFreq * mOffset, 1.0f) > 0.5f) ? -1 : 1;
 					mOffset++;
 				}
 				break;
 			case Basicwave::TRIANGLE:
 				for (i = 0; i < aSamples; i++)
 				{
-					aBuffer[i] = abs(0.5 - fmod(mParent->mFreq * mOffset, 1)) * 4 - 1;
+					aBuffer[i] = abs(0.5f - fmod(mParent->mFreq * mOffset, 1)) * 4 - 1;
 					mOffset++;
 				}
 				break;
